utils: Extend the number of bits of argument value by lsmash_bits_put().

+++ utils.c
@@ -383,8 +383,7 @@ static inline uint8_t lsmash_bits_mask_lsb8( uint32_t value, uint32_t width )
     return (uint8_t)( value & ~( ~0U << width ) );
 }
 
-/* We can change value's type to unsigned int for 64-bit operation if needed. */
-void lsmash_bits_put( lsmash_bits_t *bits, uint32_t width, uint32_t value )
+void lsmash_bits_put( lsmash_bits_t *bits, uint32_t width, uint64_t value )
 {
     debug_if( !bits || !width )
         return;
